Configuracion script .sql para la creacion de base de datos, en las subida de la app

+++ backend032/src/main/java/com/misiontic/backend032/modelos/Agencias.java
@@ -9,17 +9,17 @@ import javax.persistence.Table;
 
 // Los arroba son las anotaciones para SprintgBoot
 @Entity
-@Table(name="agencias")
+@Table(name = "agencias")
 
 public class Agencias {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int idAgencia;
 
-    @Column(name="nombreAgencia", nullable = false)
+    @Column(name = "nombreAgencia", nullable = false)
     private String nombreAgencia;
-    
-    @Column(name="direccionAgencia", nullable = false)
+
+    @Column(name = "direccionAgencia", nullable = false)
     private String direccionAgencia;
 
     public Agencias() {
@@ -28,18 +28,20 @@ public class Agencias {
     public Agencias(String nombreAgencia, String direccionAgencia) {
         this.nombreAgencia = nombreAgencia;
         this.direccionAgencia = direccionAgencia;
-        }
-        
+    }
+
     public long getIdAgencia() {
         return idAgencia;
     }
+
     public void setIdAgencia(long idAgencia) {
-        this.idAgencia = idAgencia;
+        this.idAgencia = (int) idAgencia;
     }
 
     public String getNombreAgencia() {
         return nombreAgencia;
     }
+
     public void setNombreAgencia(String nombreAgencia) {
         this.nombreAgencia = nombreAgencia;
     }
@@ -47,6 +49,7 @@ public class Agencias {
     public String getDireccionAgencia() {
         return direccionAgencia;
     }
+
     public void setDireccionAgencia(String direccionAgencia) {
         this.direccionAgencia = direccionAgencia;
     }
